updated ingredient quantity unit, change from int to double

+++ BackEnd/src/main/java/com/thegreatestteam/backend/model/Food.java
@@ -13,7 +13,7 @@ import java.util.Map;
 
 @Document("Food")
 public class Food extends Item {
-    private Map<String, Integer> components;
+    private Map<String, Double> components;
     private String type;
     private Binary image;
 
@@ -22,7 +22,7 @@ public class Food extends Item {
     private String description;
     private int kiloJoule;
 
-    public Food(String name, double price , String type, int kiloJoule, Map<String, Integer> components,Binary image, String description){
+    public Food(String name, double price , String type, int kiloJoule, Map<String, Double> components,Binary image, String description){
         super(name, price);
         this.type = type;
         this.kiloJoule = kiloJoule;
@@ -40,11 +40,11 @@ public class Food extends Item {
         this.kiloJoule = kiloJoule;
     }
 
-    public Map<String, Integer> getComponents() {
+    public Map<String, Double> getComponents() {
         return components;
     }
 
-    public void setComponents(Map<String, Integer> components) {
+    public void setComponents(Map<String, Double> components) {
         this.components = components;
     }
